Refactor Employee into Worker
