+++ src/auto/agents/mod.rs
@@ -1,3 +1,3 @@
-pub mod employee;
+pub mod worker;
 pub mod processing;
 pub mod context;
