+++ src/auto/agents/worker/tools.rs
